💄 NavBar - 뒤로가기/취소 로직 수정
