+++ src/components/common/TopNavigation/atoms/NavBack.tsx
@@ -1,20 +1,25 @@
 import Image from 'next/image';
-import Link from 'next/link';
+import { useRouter } from 'next/router';
 import { ReactElement } from 'react';
 import * as S from './styled';
 
 const NavBack = (): ReactElement => {
+  const router = useRouter();
   return (
-    <Link href='/'>
-      <S.NavItemContainer width={24} height={24}>
-        <Image
-          src='/icons/icon__nav-back.svg'
-          alt='뒤로 가기'
-          width={24}
-          height={24}
-        />
-      </S.NavItemContainer>
-    </Link>
+    <S.NavItemContainer
+      width={24}
+      height={24}
+      onClick={() => {
+        router.back();
+      }}
+    >
+      <Image
+        src='/icons/icon__nav-back.svg'
+        alt='뒤로 가기'
+        width={24}
+        height={24}
+      />
+    </S.NavItemContainer>
   );
 };
