+++ src/components/common/TopNavigation/atoms/NavCancel.tsx
@@ -1,17 +1,12 @@
-import { useRouter } from 'next/router';
+import Link from 'next/link';
 import { ReactElement } from 'react';
 import * as S from './styled';
 
 const NavCancel = (): ReactElement => {
-  const router = useRouter();
   return (
-    <S.NavItemContainer
-      onClick={() => {
-        router.back();
-      }}
-    >
+    <Link href='/'>
       <S.NavCancelText>취소</S.NavCancelText>
-    </S.NavItemContainer>
+    </Link>
   );
 };
